+++ src/configs/networks/types/ROP.js
@@ -1,3 +1,6 @@
+import tokens from '@/tokens/tokens-rop.json'
+import contracts from '@/contracts/contract-abi-rop.json'
+
 export default {
   name: 'ROP',
   name_long: 'Ropsten',
@@ -5,6 +8,6 @@ export default {
   blockExplorerTX: 'https://ropsten.etherscan.io/tx/[[txHash]]',
   blockExplorerAddr: 'https://ropsten.etherscan.io/address/[[address]]',
   chainID: 3,
-  tokens: [],
-  contracts: []
+  tokens: tokens,
+  contracts: contracts
 }