+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+### Feature
+- Feature/verify message [#52](https://github.com/MyEtherWallet/MyEtherWallet/pull/52)
+
 ### UI/UX
 - Add Create Wallet warning page [#46](https://github.com/MyEtherWallet/MyEtherWallet/pull/46)
+
+### Devop
+- Add Network logo and change dot colors [#49](https://github.com/MyEtherWallet/MyEtherWallet/pull/49)
+- Fetch tokens and contract abi [#41](https://github.com/MyEtherWallet/MyEtherWallet/pull/41)
 
 ### Release v0.0.2
 ### Feature